requirements.txt
anthropic==0.39.0
ipython==8.17.2
loguru==0.7.2
nbformat==5.10.4
numpy==1.26.4
openai==1.54.3
# pandas==1.5.3
pandas==2.2.2
protobuf==5.28.3
pytz==2022.7
PyYAML==6.0.2
tiktoken==0.7.0
boto3==1.38.2
botocore==1.38.2